feat: added build for dev
