+++ package.json
@@ -1,77 +1,80 @@
 {
-  "name": "modal-edy",
-  "version": "1.0.0",
-  "description": "micro modal",
-  "main": "./dist/index.js",
-  "module": "./dist/index.es.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/eshellster/rollup-Modal_library-Learn.git"
-  },
-  "keywords": [
-    "javascript",
-    "modal",
-    "dialog",
-    "popup"
-  ],
-  "author": "eshell",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/eshellster/rollup-Modal_library-Learn/issues"
-  },
-  "homepage": "https://github.com/eshellster/rollup-Modal_library-Learn#readme",
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged",
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
+    "name": "modal-edy",
+    "version": "1.0.0",
+    "description": "micro modal",
+    "main": "./dist/index.js",
+    "module": "./dist/index.es.js",
+    "types": "./dist/index.d.ts",
+    "files": [
+        "dist"
+    ],
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/eshellster/rollup-Modal_library-Learn.git"
+    },
+    "keywords": [
+        "javascript",
+        "modal",
+        "dialog",
+        "popup"
+    ],
+    "author": "eshell",
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/eshellster/rollup-Modal_library-Learn/issues"
+    },
+    "homepage": "https://github.com/eshellster/rollup-Modal_library-Learn#readme",
+    "husky": {
+        "hooks": {
+            "pre-commit": "lint-staged",
+            "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
+        }
+    },
+    "lint-staged": {
+        "src/**/*.{js,ts}": [
+            "yarn lint"
+        ]
+    },
+    "scripts": {
+        "prettier": "prettier '**/*.{js,ts,html}' --ignore-path ./.prettierignore",
+        "test": "echo \"Error: no test specified\" && exit 1",
+        "types": "tsc --noEmit",
+        "lint": "eslint '*/**/*.{js,ts}'",
+        "finish": "yarn lint && yarn types",
+        "build": "yarn build:clean && yarn build:lib",
+        "build:clean": "rimraf dist",
+        "build:lib": "rollup --config config/rollup/rollup.config.lib.js",
+        "build:dev": "rollup --config config/rollup/rollup.config.dev.js -w"
+    },
+    "devDependencies": {
+        "@babel/core": "^7.12.10",
+        "@babel/preset-env": "^7.12.11",
+        "@babel/preset-typescript": "^7.12.7",
+        "@commitlint/cli": "^11.0.0",
+        "@commitlint/config-conventional": "^11.0.0",
+        "@rollup/plugin-babel": "^5.2.2",
+        "@rollup/plugin-commonjs": "^17.0.0",
+        "@rollup/plugin-node-resolve": "^11.0.1",
+        "@semantic-release/changelog": "^5.0.1",
+        "@semantic-release/commit-analyzer": "^8.0.1",
+        "@semantic-release/git": "^9.0.0",
+        "@semantic-release/npm": "^7.0.9",
+        "@semantic-release/release-notes-generator": "^9.0.1",
+        "@types/eslint": "^7.2.6",
+        "@typescript-eslint/eslint-plugin": "^4.10.0",
+        "@typescript-eslint/parser": "^4.10.0",
+        "eslint": "^7.16.0",
+        "eslint-config-prettier": "^7.1.0",
+        "eslint-plugin-jsdoc": "^30.7.9",
+        "eslint-plugin-prettier": "^3.3.0",
+        "husky": "^4.3.6",
+        "lint-staged": "^10.5.3",
+        "prettier": "^2.2.1",
+        "rollup": "^2.35.1",
+        "rollup-plugin-livereload": "^2.0.0",
+        "rollup-plugin-serve": "^1.1.0",
+        "rollup-plugin-terser": "^7.0.2",
+        "rollup-plugin-typescript2": "^0.29.0",
+        "typescript": "^4.1.3"
     }
-  },
-  "lint-staged": {
-    "src/**/*.{js,ts}": [
-      "yarn lint"
-    ]
-  },
-  "scripts": {
-    "prettier": "prettier '**/*.{js,ts,html}' --ignore-path ./.prettierignore",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "types": "tsc --noEmit",
-    "lint": "eslint '*/**/*.{js,ts}'",
-    "finish": "yarn lint && yarn types",
-    "build": "yarn build:clean && yarn build:lib",
-    "build:clean": "rimraf dist",
-    "build:lib": "rollup -c"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.12.10",
-    "@babel/preset-env": "^7.12.11",
-    "@babel/preset-typescript": "^7.12.7",
-    "@commitlint/cli": "^11.0.0",
-    "@commitlint/config-conventional": "^11.0.0",
-    "@rollup/plugin-babel": "^5.2.2",
-    "@rollup/plugin-commonjs": "^17.0.0",
-    "@rollup/plugin-node-resolve": "^11.0.1",
-    "@semantic-release/changelog": "^5.0.1",
-    "@semantic-release/commit-analyzer": "^8.0.1",
-    "@semantic-release/git": "^9.0.0",
-    "@semantic-release/npm": "^7.0.9",
-    "@semantic-release/release-notes-generator": "^9.0.1",
-    "@types/eslint": "^7.2.6",
-    "@typescript-eslint/eslint-plugin": "^4.10.0",
-    "@typescript-eslint/parser": "^4.10.0",
-    "eslint": "^7.16.0",
-    "eslint-config-prettier": "^7.1.0",
-    "eslint-plugin-jsdoc": "^30.7.9",
-    "eslint-plugin-prettier": "^3.3.0",
-    "husky": "^4.3.6",
-    "lint-staged": "^10.5.3",
-    "prettier": "^2.2.1",
-    "rollup": "^2.35.1",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.29.0",
-    "typescript": "^4.1.3"
-  }
 }
